# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Resonance API",
    "version": "1.0.0",
    "description": "Cookie-authenticated REST API. Fetch /auth/me for a CSRF token before writes. Admin endpoints require the admin role."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/admin/imports": {
      "get": {
        "summary": "List import jobs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "summary": "Queue video links or public playlist discovery",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "string"
                  }
                },
                "required": ["input"]
              }
            }
          }
        }
      }
    },
    "/admin/imports/{id}": {
      "get": {
        "summary": "Import progress, suggestions, editable rows and failures",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/admin/imports/{id}/retry": {
      "post": {
        "summary": "Resume failed import",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/imports/{id}/publish": {
      "post": {
        "summary": "Publish reviewed rows atomically per row",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "videoIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "maxItems": 200
                  }
                },
                "required": ["videoIds"]
              }
            }
          }
        }
      }
    },
    "/admin/imports/{id}/rows/{videoId}": {
      "patch": {
        "summary": "Save reviewed music metadata",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "videoId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportDraft"
              }
            }
          }
        }
      }
    },
    "/admin/imports/{id}/rows/{videoId}/match": {
      "post": {
        "summary": "Resolve a selected MusicBrainz recording and release; return genres and track positions",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failure"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or CSRF check failed"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "videoId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["recordingId"],
                "properties": {
                  "recordingId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "releaseId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "resonance.sid"
      }
    },
...
  }
}